-fixed the head looking weird -fixed potential name going to next line -changed getting to location to be using google maps (now it works) -added the scroll view -changed location into having double latitude and longitude

+++ QRHunter/app/src/main/java/com/example/qrhunter/data/model/Location.java
@@ -6,12 +6,12 @@ import java.util.ArrayList;
  * A class representing a location
  */
 public class Location {
-    public int latitude;
-    public int longitude;
+    public double latitude;
+    public double longitude;
 
     public ArrayList<String> photos;
 
-    public Location(int latitude, int longitude, ArrayList<String> photos) {
+    public Location(double latitude, double longitude, ArrayList<String> photos) {
         this.latitude = latitude;
         this.longitude = longitude;
         this.photos = photos;
